@@ -4,36 +4,80 @@
   "description": "",
   "main": "index.js",
   "dependencies": {
+    "autobind-decorator": "^1.3.4",
+    "babel-plugin-transform-class-properties": "^6.19.0",
+    "babel-plugin-transform-decorators-legacy": "^1.3.4",
+    "babel-plugin-transform-object-rest-spread": "^6.16.0",
     "babel-preset-es2015": "^6.14.0",
     "babel-preset-react": "^6.11.1",
+    "babel-preset-stage-2": "^6.18.0",
     "babelify": "^7.3.0",
+    "bluebird": "^3.4.6",
     "bower": "^1.7.9",
     "browserify": "^13.1.0",
+    "class-autobind": "^0.1.4",
+    "classnames": "^2.2.5",
+    "cldr-data": "^30.0.2",
     "css-loader": "^0.25.0",
     "formsy-react": "^0.18.1",
-    "jquery": "^3.1.0",
+    "globalize": "^1.2.1",
+    "jquery": "^3.1.1",
+    "jquery-ui": "^1.12.1",
+    "moment": "^2.16.0",
+    "newforms": "^0.13.2",
     "node": "0.0.0",
     "object-assign": "^4.1.0",
-    "react": "^15.3.1",
-    "react-dom": "^15.3.1",
+    "opentok-react": "^0.3.0",
+    "pretty-checkbox": "^2.1.0",
+    "rc-time-picker": "^2.2.1",
+    "react": "^15.4.1",
+    "react-big-calendar": "^0.11.1",
+    "react-currency-input": "^1.1.1",
+    "react-datepicker": "^0.37.0",
+    "react-datetime": "^2.6.0",
+    "react-dimensions": "^1.3.0",
+    "react-dom": "^15.4.1",
+    "react-dropzone": "^3.9.0",
+    "react-dropzone-s3-uploader": "^0.10.0",
+    "react-global": "^0.1.8",
+    "react-iframe": "0.0.4",
+    "react-maskedinput": "^3.3.4",
+    "react-measure": "^1.4.5",
+    "react-modal": "^1.5.2",
+    "react-notification-system": "^0.2.11",
+    "react-render-html": "^0.1.6",
+    "react-rotating-text": "^1.1.2",
     "react-router": "^2.8.1",
+    "react-select": "^1.0.0-rc.2",
     "react-semantic-ui": "^0.2.0",
+    "react-share": "^1.11.0",
+    "react-text-truncate": "^0.8.3",
+    "react-tinymce": "^0.5.1",
+    "react-tinymce-input": "^1.2.1",
+    "react-validated-input": "0.0.10",
     "remarkable": "^1.6.2",
-    "style-loader": "^0.13.1"
+    "searchkit": "^0.10.1",
+    "semantic-ui-daterangepicker": "^1.3.21",
+    "semantic-ui-sidebar": "^2.2.3",
+    "style-loader": "^0.13.1",
+    "twilio-video": "^1.0.0-beta3",
+    "validator": "^6.1.0"
   },
   "devDependencies": {
     "babel": "^6.5.2",
     "babel-core": "^6.14.0",
     "babel-loader": "^6.2.5",
     "babel-preset-react-hmre": "^1.1.1",
+    "jsx-control-statements": "^3.1.5",
     "react": "^15.3.1",
-    "react-hot-loader": "^3.0.0-beta.4",
+    "react-hot-loader": "^1.3.1",
     "webpack": "^1.13.2",
     "webpack-bundle-tracker": "0.0.93"
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "author": "",
-    "license": "ISC"
+    "license": "ISC",
+    "start": "node server.js"
   }
 }
