@@ -1,8 +1,22 @@
-var React = require('react')
-var ReactDOM = require('react-dom')
+var React = require('react');
+var ReactDOM = require('react-dom');
 var $  = require('jquery');
-global.rsui = require('react-semantic-ui')
-var forms = require('newforms')
+global.rsui = require('react-semantic-ui');
+var forms = require('newforms');
+import { Router, Route, Link, browserHistory, hashHistory } from 'react-router'
+var ObjectPage = require('./step');
+import autobind from 'class-autobind'
+import DropzoneS3Uploader from 'react-dropzone-s3-uploader'
+
+
+
+var Global = require('react-global');
+
+<Global values={{
+  isSidebarVisible: 'false'
+}} />
+
+var theServer = 'https://192.168.1.156:8000/'
 
 
 
@@ -21,26 +35,49 @@ var ObjectListAndUpdate = React.createClass({
           }.bind(this)
         });
       },
+    handleFormSubmit: function (formData) {
+    $.ajax({
+        url: this.props.url,
+        dataType: 'json',
+        contentType: "application/json",
+
+        type: 'POST',
+        data: formData,
+
+        success: function(data) {
+            this.setState({data: data});
+        }.bind(this),
+        error: function(xhr, status, err) {
+            console.error(this.props.url, status, err.toString());
+        }.bind(this)
+        });
+  },
 
     getInitialState: function() {
         return {data: []};
     },
 
     componentDidMount: function() {
-        this.loadObjectsFromServer()
-        //setInterval(this.loadObjectsFromServer, this.props.pollInterval);
+        this.loadObjectsFromServer();
+        var intervalID = setInterval(this.loadObjectsFromServer, 2000);
+        this.setState({intervalID:intervalID});
 
         var self = this;
     },
+
+    componentWillUnmount: function() {
+   // use intervalId from the state to clear the interval
+    clearInterval(this.state.intervalId);
+    },
     render:function() {
 
-        var model = this.props.model
+        var model = this.props.model;
 
         return (
-            <div>
-                <FormAction pageHeadingLabel={this.props.pageHeadingLabel} actionButtonLabel={this.props.actionButtonLabel} actionFormRef={this.props.actionFormRef} modelForm={this.props.modelForm}/>
+<div>
+                <FormAction onFormSubmit={this.handleFormSubmit} pageHeadingLabel={this.props.pageHeadingLabel} actionButtonLabel={this.props.actionButtonLabel} actionFormRef={this.props.actionFormRef} modelForm={this.props.modelForm}/>
 
-                <ObjectList data={this.state.data}  pollInterval={2000} />
+                <ObjectList data={this.state.data}  />
 
             </div>
         );
@@ -49,95 +86,264 @@ var ObjectListAndUpdate = React.createClass({
 
 
 var ObjectList = React.createClass({
+    toggle: function(planId) {
+        $(this.refs["planFormRef_" + planId]).slideToggle()
+    },
 
-    render: function() {
+    editPlan: function(planId) {
+        $(this.refs["planFormRef_" + planId]).slideToggle();
 
-        if (this.props.data) {
-            var objectNodes = this.props.data.map(function (objectData) {
-                return (
-                    <div className="ui fluid column" key={objectData.id}>{objectData.title}</div>
-                )
-            });
-            console.log(objectNodes)
-        };
+    },
 
-
-
-
-        return (
-            <div>
-                {objectNodes}
-            </div>
-            )
-}});
-
-var FormAction = React.createClass({
     componentDidMount: function() {
-        $(this.refs[this.props.actionFormRef]).hide()
-    },
-    toggle: function() {
-        $(this.refs[this.props.actionFormRef]).slideToggle()
-        $(this.refs['clickToToggleButton']).hide()
+        $(this.refs[this.props.actionFormRef]).slideToggle();
+
+
     },
 
-    handleSubmit: function(e) {
+    componentDidUpdate: function() {
+        $(".formForHiding").hide();
+
+    },
+
+
+    hideForm: function(planId) {
+        console.log("hidden");
+        $(this.refs["planFormRef_" + planId]).hide();
+
+    },
+    clearPage: function(plan_id) {
+        $(".fullPageDiv").slideToggle("slow", function () {
+        hashHistory.push('/plans/' + plan_id + '/steps')
+            });
+
+    },
+
+    handleSubmit2: function(e) {
+        console.log("handleSubmit called");
         e.preventDefault();
         var title = this.state.title;
-        var deadline = this.state.deadline;
-        var why = this.state.why;
-        var votes = this.state.votes;
         var viewableBy = this.state.viewableBy;
-        var image = this.state.image;
         var description = this.state.description;
 
 
         if (!description || !title ) {
         return;
         }
-        this.props.onGoalSubmit({title: title, deadline:deadline, why:why, votes:votes, viewableBy:viewableBy, image:image, description: description});
-        this.setState({title: '',  deadline: '', why: '', votes:0, image:'', description: '', viewableBy: 'Only me', });
+        this.props.onFormSubmit({title: title, viewableBy:viewableBy, description: description});
+        this.setState({title: '',   description: '', viewableBy: 'Only me', });
     },
 
     render: function() {
-        console.log("Modelform is" + this.props.modelForm)
-        switch(this.props.modelForm) {
+        var theForm = new PlanForm();
+
+        if (this.props.data) {
+            var objectNodes = this.props.data.map(function (objectData) {
+
+                return (
+                    <div className="column " key={objectData.id} >
+                        <div className="ui top attached green button" onClick={this.clearPage.bind(this, objectData.id)}>
+                                  Plan
+                                </div>
+                    <div className="ui segment noBottomMargin noTopMargin">
+
+                            <div className="row">
+                                <button className="right floated ui small green button"  onClick={this.editPlan.bind(this, objectData.id)}>Edit Plan</button>
+                                <h3>{objectData.title}</h3>
+                            </div>
+                            <div className="fluid row">Start Date (just as reference): {objectData.startDate}</div>
+                            <div className="fluid row">Length: {objectData.scheduleLength} </div>
+                            <div className="fluid row">{objectData.description}</div>
+
+                        <div className="two wide column"><Link to={`/#/plans/${objectData.id}/steps`}>
+
+</Link>
+                    </div>
+                    <div ref={`planFormRef_${objectData.id}`} className="sixteen wide row formForHiding">
+                    <div ref="planForm1">
+                        <div className="ui form">
+                            <form onSubmit={this.handleSubmit}>
+                                <forms.RenderForm form={theForm} enctype="multipart/form-data" ref="theFormRef"/>
+
+                                <div className="ui grid">
+                                    <div className="ui row">&nbsp;</div>
+
+                                    <div className="ui row">
+
+
+                                        <div className="eight wide column">
+                                            <button className="ui fluid button" onClick={this.toggle.bind(this, objectData.id)}>Cancel</button>
+                                        </div>
+                                        <div className="eight wide column">
+                                            <button className="ui primary fluid button" type="submit">Save</button>
+                                        </div>
+
+
+                                    </div>
+                                                                        <div className="ui row">&nbsp;</div>
+
+                                </div>
+                            </form>
+                        </div>
+                    </div>
+                </div>
+
+                        </div>
+
+                        <div className="ui bottom attached purple button" onClick={this.clearPage.bind(this, objectData.id)}>
+                                Steps
+                                </div>
+                        </div>
+
+
+
+
+
+
+
+
+
+                )
+            }.bind(this));
+        }
+        return (
+            //<div className="ui divided link items">
+            <div className="ui three column grid">
+                {objectNodes}
+            </div>
+            )
+}});
+
+
+
+var FormAction = React.createClass({
+    componentDidMount: function() {
+        $(this.refs[this.props.actionFormRef]).hide()
+    },
+    toggle: function() {
+        $(this.refs[this.props.actionFormRef]).slideToggle();
+        $(this.refs['clickToToggleButton']).toggle()
+    },
+
+    handleSubmit: function(e) {
+        console.log("handleSubmit called");
+        e.preventDefault();
+        var title = this.state.title;
+        var viewableBy = this.state.viewableBy;
+        var description = this.state.description;
+
+
+        if (!description || !title ) {
+        return;
+        }
+        this.props.onFormSubmit({title: title, viewableBy:viewableBy, description: description});
+        this.setState({title: '',   description: '', viewableBy: 'Only me', });
+    },
+
+
+    render: function() {
+        var buttonColor;
+         switch(this.props.modelForm) {
             case "PlanForm":
                 var theForm = new PlanForm();
+                 buttonColor = "green";
                 break;
             case "StepForm":
                 var theForm = new StepForm();
+                 buttonColor = "purple";
+
                 break;
             default:
+                 buttonColor = "blue";
+
                 break;
         }
         return (
             <div>
-            <div className="ui grid">
-                <div className="ui four wide column header"><h1>{this.props.pageHeadingLabel}</h1></div>
-                <div className="ui right floated four wide column">
-                    <button className="ui right floated primary large fluid button" ref="clickToToggleButton" onClick={this.toggle}>{this.props.actionButtonLabel}</button>
+            <div className="ui three column grid">
+                <div className="ui column header"><h1>{this.props.pageHeadingLabel}</h1></div>
+                <div className="ui right floated column">
+                    <button className={`ui right floated  large fluid button ${buttonColor}`} ref="clickToToggleButton" onClick={this.toggle}>{this.props.actionButtonLabel}</button>
                 </div>
             </div>
             <div ref={`${this.props.actionFormRef}`}><div className="ui form"><form onSubmit={this.handleSubmit}>
         <forms.RenderForm form={theForm} enctype="multipart/form-data" ref="theFormRef" />
-                </form></div>
-                <div className="ui grid">
+
+                <div className="ui three column grid">
                                         <div className="ui row">&nbsp;</div>
 
                     <div className="ui row">
-                        <div className="ui eight wide column">&nbsp;</div>
+                        <div className="ui column">&nbsp;</div>
 
 
-                                <div className="ui four wide column"><button className="ui fluid button">Cancel</button></div>
-                                <div className="ui  four wide column"><button type="submit" type="submit" form={theForm} className="ui primary fluid button">Save</button></div>
+                                <div className="ui column"><button className="ui fluid button" onClick={this.toggle}>Cancel</button></div>
+                                <div className="ui  column"><button type="submit" className="ui primary fluid button">Save</button></div>
 
 
                     </div></div>
-
-            </div></div>
+</form></div>
+            </div>
+            </div>
         )
     }
-})
+});
+
+var DynamicSelectButton2 = React.createClass({
+    componentDidMount: function() {
+        var self = this;
+    },
+
+    getInitialState: function() {
+        return {
+            value: this.props.initialValue,
+        }
+    },
+
+  render: function () {
+      var htmlToRender = "<div className='field'><label htmlFor='" + this.props.id + "'>" + this.props.label + "</label>";
+      htmlToRender += "<select id='" + this.props.id + "' name='" + this.props.initialValue + "' >" ;
+
+      for(var currentItem in this.props.items) {
+          if (this.props.initialValue == currentItem){
+              htmlToRender += "<option selected='selected' value='";
+          }
+          else {
+              htmlToRender += "<option value='";
+          }
+          htmlToRender += "<option value='";
+          htmlToRender += String(this.props[currentItem]);
+          htmlToRender += "'>";
+          htmlToRender += String(currentItem);
+          htmlToRender += "</option>";
+
+      }
+      htmlToRender += "</select>";
+    htmlToRender += "</div>";
+    return (
+<div dangerouslySetInnerHTML={{__html: htmlToRender}} />
+      );
+
+  }
+});
+
+var TimePicker = React.createClass({
+    render: function(){
+        return <Datetime
+            renderDay={ this.renderDay }
+            renderMonth={ this.renderMonth }
+            renderYear={ this.renderYear }
+        />;
+    },
+    renderDay: function( props, currentDate, selectedDate ){
+        return "";
+    },
+    renderMonth: function( props, month, year, selectedDate){
+        return "";
+    },
+    renderYear: function( props, year, selectedDate ){
+        return "";
+    }
+});
 
 var PlanForm = forms.Form.extend({
 
@@ -146,30 +352,135 @@ var PlanForm = forms.Form.extend({
         description:forms.CharField({widget: forms.Textarea()}),
         viewableBy:forms.ChoiceField({choices:["Only me": "ONLY_ME", "Just people I've shared this goal with": "SHARED", "Just my Pros": "MY_PROS", "All Pros": "ALL_PROS", "Everyone": "EVERYONE"]}),
         startDate: forms.CharField({widget: forms.DateTimeInput()}),
-        endDate: forms.CharField({widget: forms.DateTimeInput()}),
-        duration: forms.IntegerField(),
-        durationMetric: forms.ChoiceField({choices:["weeks": "weeks", "days":"days", "months":"months", "years":"years"]}),
-        calendar:forms.FileField({widget: forms.FileInput({attrs: {className: 'ui button', css:'opacity:0;'}}), label: "Import Calendar (.ics) File", css:"ui button"}),
+            endDate: forms.CharField({widget: forms.DateTimeInput()}),
+        scheduleLength: forms.IntegerField(),
+        //calendar:forms.FileField({widget: forms.FileInput({attrs: {className: 'ui button', css:'opacity:0;'}}), label: "Import Calendar (.ics) File", css:"ui button"}),
+
 
     })
 
 
-var StepForm = forms.Form.extend({
-            rowCssClass: 'field',
+var Sidebar = React.createClass({
+    render: function() {
 
-        title: forms.CharField(),
-        description: forms.CharField(),
-        frequency: forms.MultipleChoiceField(),
-        onMonday:forms.BooleanField(),
-        onTuesday:forms.BooleanField(),
-        onWednesday:forms.BooleanField(),
-        onThursday:forms.BooleanField(),
-        onFriday:forms.BooleanField(),
-        onSaturday:forms.BooleanField(),
-        onSunday:forms.BooleanField(),
-        startDate: forms.MultipleChoiceField(),
-        endDate: forms.MultipleChoiceField(),
-    })
+        if (Global.get('sidebarVisible') == 'true') {
+            var isSidebarVisible = "";
+        } else {
+            var isSidebarVisible = "";
+    }
+        return (
+
+            <div className={`ui left vertical inverted labeled icon ${isSidebarVisible} sidebar menu`}>
+                            <div className="sidebar-spacer">&nbsp;</div>
+
+                <Link className="item" to="/">
+        <i className="home icon"></i>
+        Home
+    </Link>
+    <Link className="item" to="/goals">
+        <i className="block layout icon"></i>
+        Goals
+    </Link>
+    <a className="item">
+        <i className="smile icon"></i>
+        Friends
+    </a>
+    <a className="item">
+        <i className="calendar icon"></i>
+        History
+    </a>
+    <a className="item">
+        <i className="mail icon"></i>
+        Messages
+    </a>
+    <a className="item">
+        <i className="chat icon"></i>
+        Discussions
+    </a>
+    <a className="item">
+        <i className="trophy icon"></i>
+        Achievements
+    </a>
+    <a className="item">
+        <i className="shop icon"></i>
+        Store
+    </a>
+    <a className="item">
+        <i className="settings icon"></i>
+        Settings
+    </a>
+  </div>
+        )
+    }
+});
 
 
-module.exports = {FormAction, ObjectList, ObjectListAndUpdate}
+const dropzoneS3Style = {
+    height: 200,
+    border: 'dashed 2px #999',
+    borderRadius: 5,
+    position: 'relative',
+    cursor: 'pointer',
+  }
+
+  const uploaderProps = {
+    dropzoneS3Style,
+    maxFileSize: 1024 * 1024 * 50,
+    server: theServer,
+    s3Url: 'https://kiterope.s3.amazonaws.com/images',
+    signingUrlQueryParams: {uploadType: 'avatar'},
+      uploadRequestHeaders: {'x-amz-acl': 'public-read','Access-Control-Allow-Origin':'*' },
+      signingUrl: "signS3Upload",
+  }
+
+const s3ImageUrl = "https://kiterope.s3.amazonaws.com:443/"
+
+export class ImageUploader extends React.Component {
+    constructor(props) {
+        super(props)
+        autobind(this)
+        this.state = {
+            image:"",
+        }
+    }
+
+    componentDidMount() {
+        this.setState({
+            defaultImage: this.props.defaultImage,
+        })
+    }
+
+    handleFinishedUpload (value) {
+            var fullUrl = value.signedUrl;
+            var urlForDatabase = fullUrl.split("?")[0];
+            urlForDatabase = urlForDatabase.replace(s3ImageUrl, "");
+            this.setState({image: urlForDatabase});
+    }
+
+    render() {
+                    var theImage = this.state.image
+
+                    var theFilename = theImage.replace("https://kiterope.s3.amazonaws.com:443/images/", "");
+
+        return (
+            <div className="field">
+            <label htmlFor="id_image">{this.props.label}</label>
+            <DropzoneS3Uploader filename={theFilename} onFinish={this.handleFinishedUpload} {...uploaderProps} />
+                </div>
+        )
+    }
+}
+
+
+
+
+
+module.exports = {
+    FormAction,
+    ObjectList,
+    ObjectListAndUpdate,
+    TimePicker,
+    Sidebar,
+    DynamicSelectButton2,
+    ImageUploader,
+}
