@@ -1,43 +1,117 @@
-var React = require('react')
-var ReactDOM = require('react-dom')
+var React = require('react');
+var ReactDOM = require('react-dom');
 var $  = require('jquery');
-global.rsui = require('react-semantic-ui')
-var forms = require('newforms')
-import {ObjectList, ObjectListAndUpdate, FormAction } from './base'
+global.rsui = require('react-semantic-ui');
+var forms = require('newforms');
+import {ObjectList, ObjectListAndUpdate, FormAction, Sidebar} from './base';
+import {ObjectPage, PlanHeader} from './step';
+import { Router, Route, Link, browserHistory, hashHistory } from 'react-router';
 
-{/*
-ObjectCreationPage
-    ObjectListAndUpdate()
-        PageHeadingBar('Plans', 'Add Plan' PlanForm)
-        ObjectList('Plan')
-        PlanContainer
-            FormRenderer
-                PlanForm
-                    CharField('title');
-                    CharField('description')
-                    Charfield('viewableBy', withChoices="VIEWABLE_CHOICES")
-                    DateTimeField('duration')
-                    FormSubmissionActionButtons(planForm)
-            StepContainer
-                ObjectListAndUpdate()
-                    PageHeadingBar('Steps', 'Add Step': StepForm)
-                    ObjectListWithSubCreation('Step')
-                    StepForm
-                        CharField('description')
-                        CharField('frequency')
-                        WeeklyBooleanSet('onSunday', 'onMonday', etc.)
-                        FormSubmissionActionButtons(stepForm)
-                        */}
+
+
+$.ajaxSetup({
+    beforeSend: function(xhr) {
+        xhr.setRequestHeader('X-CSRFToken', getCookie('csrftoken'));
+    }
+});
+
+
+var theServer = 'https://192.168.1.156:8000/'
+
 
 var ObjectCreationPage = React.createClass({
+    componentWillMount: function() {
+
+    },
+    componentDidMount: function() {
+        $(".fullPageDiv").hide();
+        $(".fullPageDiv").slideToggle();
+
+
+    },
+
     render:function() {
-        return (
-            <div>
-            <div className="spacer">&nbsp;</div>
-            <div className="ui alert"></div>
-                <ObjectListAndUpdate url="http://127.0.0.1:8000/api/plans/" pageHeadingLabel="Plans" actionButtonLabel="Add Plan" actionFormRef="planForm" modelForm="PlanForm" />
-            </div>
-        );
+
+        if (this.props.params.plan_id) {
+            return (
+
+                <div className="fullPageDiv ">
+                                        <div className="ui page container">
+
+                <div className="spacer">&nbsp;</div>
+                <div className="ui alert"></div>
+                    <div className="ui large breadcrumb">
+                         <Link to={`/`}><div className="section">Home</div></Link>
+                            <i className="right chevron icon divider"></i>
+                            <Link to={`/goals`}><div className="active section">My Goals</div></Link>
+                        <i className="right chevron icon divider"></i>
+                            <Link to={`/plans/${this.props.params.plan_id}/`}><div className="active section">Plan Detail</div></Link>
+                    </div>
+            <div>&nbsp;</div>
+                    <PlanHeader url={`${theServer}api/plans/${this.props.params.plan_id}`} />
+                    <div>&nbsp;</div>
+
+                    <ObjectPage url={`${theServer}api/steps/`}
+                                                               pageHeadingLabel="Steps"
+                                                               actionButtonLabel="Add Step"
+                                                               actionFormRef="stepForm"
+                                                               modelForm="StepForm"/>
+
+
+                </div></div>
+
+            )
+        }
+        else if (this.props.params.goal_id) {
+            return (
+
+
+                <div className="fullPageDiv">
+                    <div className="ui page container">
+                <div className="spacer">&nbsp;</div>
+                <div className="ui alert"></div>
+                    <div className="ui large breadcrumb">
+                         <Link to={`/`}><div className="section">Home</div></Link>
+                            <i className="right chevron icon divider"></i>
+                            <Link to={`/goals`}><div className="active section">My Goals</div></Link>
+                        <i className="right chevron icon divider"></i>
+                            <Link to={`/goals/${this.props.params.goal_id}/plans/`}><div className="active section">Goal Detail</div></Link>
+                    </div>
+                    <div>&nbsp;</div>
+                    <GoalHeader url={`${theServer}api/goals/${this.props.params.goal_id}`} />
+                    <div>&nbsp;</div>
+
+                    <ObjectListAndUpdate url={`${theServer}api/plans/`} pageHeadingLabel="Plans" actionButtonLabel="Add Plan" actionFormRef="planForm" modelForm="PlanForm" />
+                    {/*<PlanHeader url={`http://127.0.0.1:8000/api/goals/${this.props.params.goal_id}/plans`} />*/}
+
+
+                </div>
+    </div>
+
+            )
+
+        }
+        else {
+            return (
+
+                <div className="fullPageDiv">
+                                        <div className="ui page container">
+
+                <div className="spacer">&nbsp;</div>
+                <div className="ui alert"></div>
+                    <div className="ui large breadcrumb">
+                         <Link to={`/`}><div className="section">Home</div></Link>
+                            <i className="right chevron icon divider"></i>
+                            <Link to={`/goals/`}><div className="active section">My Goals</div></Link>
+                        <i className="right chevron icon divider"></i>
+                            <Link to={`/plans/`}><div className="active section">My Plans</div></Link>
+                    </div>
+                    <div>&nbsp;</div>
+                    <ObjectListAndUpdate url={`${theServer}api/plans/`} pageHeadingLabel="Plans" actionButtonLabel="Add Plan" actionFormRef="planForm" modelForm="PlanForm" />
+</div></div>
+
+            );
+        }
     }
 });
 
@@ -46,8 +120,79 @@ var ObjectCreationPage = React.createClass({
 
 
 
+var GoalHeader = React.createClass({
+    loadObjectsFromServer: function () {
+        console.log(this.props.url);
+        $.ajax({
+          url: this.props.url,
+          dataType: 'json',
+          cache: false,
+          success: function(data) {
+
+                  this.setState({
+                      title:data.title,
+                      description:data.description,
+                      data: data});
 
 
+          }.bind(this),
+          error: function(xhr, status, err) {
+            console.error(this.props.url, status, err.toString());
+          }.bind(this)
+        });
+      },
+
+    getInitialState: function() {
+
+        return {data: []};
+    },
+
+    componentDidMount: function() {
+        this.loadObjectsFromServer();
+          var intervalID = setInterval(this.loadObjectsFromServer, 2000);
+        this.setState({intervalID:intervalID});
+
+        var self = this;
+    },
+
+    componentWillUnmount: function() {
+   // use intervalId from the state to clear the interval
+   clearInterval(this.state.intervalId);
+},
+
+    render: function() {
+
+        return (
+            <div>
+                        <div className="ui four wide column header"><h1>Goal</h1></div>
+<div className="ui top attached primary button" >
+                                  Goal
+                                </div>
+                <div className="ui segment noTopMargin two column grid">
+
+                <div className="four wide column">
+                    <img className="ui image" src='http://semantic-ui.com/images/avatar2/large/kristy.png'></img>
+                    </div>
+
+                        <div className="twelve wide column">
+                            <div className="row">
+                            <div className="sixteen wide column">
+                                <h1>{this.state.title}</h1></div>
+                        </div>
+
+
+                            <div className="row">
+
+                            <div className="sixteen wide column"> {this.state.description}</div>
+
+
+
+                    </div>
+                </div>
+            </div>
+                </div>
+            )
+}});
 
 
 
@@ -55,13 +200,13 @@ var ObjectCreationPage = React.createClass({
 var StepContainer = React.createClass({
     render: function() {
         return (
-        <ObjectListAndUpdate url="http://127.0.0.1:8000/api/steps/" pageHeadingLabel="Steps" actionButtonLabel="Add Step" actionFormRef="stepForm" model="step" />
+        <ObjectListAndUpdate url={`${theServer}api/steps/`} pageHeadingLabel="Steps" actionButtonLabel="Add Step" actionFormRef="stepForm" model="step" />
 
 
 
         )
     }
-})
+});
 
 
 var PlanFormAction = React.createClass({
@@ -88,14 +233,14 @@ var PlanFormAction = React.createClass({
             </div>
         )
     }
-})
+});
 
 var FormRenderer = React.createClass({
         render:function() {
             var theModel = this.props.model;
             var theRef = this.props.divReference;
             var theActionLabel = this.props.actionLabel;
-            var formClass = "ui form"
+            var formClass = "ui form";
 
             return (
                                     <div className="ui form">
@@ -111,8 +256,22 @@ var FormRenderer = React.createClass({
     });
 
 
+function getCookie(name) {
+    var cookieValue = null;
+    if (document.cookie && document.cookie != '') {
+        var cookies = document.cookie.split(';');
+        for (var i = 0; i < cookies.length; i++) {
+            var cookie = cookies[i].trim();
+            // Does this cookie string begin with the name we want?
+            if (cookie.substring(0, name.length + 1) == (name + '=')) {
+                cookieValue = decodeURIComponent(cookie.substring(name.length + 1));
+                break;
+            }
+        }
+    }
+    return cookieValue;
+}
 
 
 
-
-module.exports = ObjectCreationPage;
+module.exports = ObjectCreationPage, GoalHeader;
