@@ -1,5 +1,6 @@
 from django.contrib import admin
-from kiterope.models import Profile, User, Goal, Student, Coach, Rate, Session, Review, Update, Post, Plan, Step, Question, Answer, Interest
+from kiterope.models import Profile, User, Goal, Student, Notification, Coach, Rate, Session, Review, Update, Post, Plan, Step, Question, Answer, Interest, Participant
+from kiterope.models import StepOccurrence, PlanOccurrence, UpdateOccurrence, Metric
 admin.site.register(Profile)
 admin.site.register(Goal)
 admin.site.register(Student)
@@ -14,3 +15,14 @@ admin.site.register(Step)
 admin.site.register(Question)
 admin.site.register(Answer)
 admin.site.register(Interest)
+admin.site.register(StepOccurrence)
+admin.site.register(PlanOccurrence)
+admin.site.register(UpdateOccurrence)
+admin.site.register(Metric)
+admin.site.register(Participant)
+admin.site.register(Notification)
+
+
+
+
+
