@@ -19,13 +19,14 @@ THUMBNAIL_DEBUG = True
 CORS_ORIGIN_ALLOW_ALL = True
 
 
-
 '''VP STATICFILES_DIRS = (
     os.path.join(BASE_DIR, "static"),
     '/Users/eric/Dropbox/_syncFolder/Business/kiterope/code/kiterope/static/',
 )'''
 
-
+INSTALLED_APPS += [
+    'sslserver',
+]
 
 
 
