@@ -27,6 +27,8 @@ import oauth2_provider.views as oauth2_views
 
 
 
+
+
 import apps
 from kiterope import views
 
@@ -36,9 +38,31 @@ from rest_framework.urlpatterns import format_suffix_patterns
 
 router = routers.DefaultRouter()
 router.register(r'users', views.UserViewSet)
-router.register(r'goals', views.GoalViewSet)
 router.register(r'plans', views.PlanViewSet)
+router.register(r'updates', views.UpdateViewSet)
 router.register(r'steps', views.StepViewSet)
+router.register(r'searchQuery', views.SearchQueryViewSet, 'SearchQuery')
+router.register(r'goalEntry', views.GoalViewSet, base_name='Goal')
+router.register(r'sessions', views.SessionViewSet, base_name='Session')
+router.register(r'notifications', views.NotificationViewSet, base_name='Notification')
+
+
+router.register(r'^plans/(?P<plan_id>\w+)/steps', views.StepViewSet, base_name='Step')
+router.register(r'^period/(?P<periodRangeStart>[\w\-]+)/(?P<periodRangeEnd>[\w\-]+)', views.StepOccurrenceViewSet, base_name="StepOccurrence")
+#router.register(r'period', views.GoalViewSet, base_name='Goal')
+router.register(r'^steps/(?P<step_id>\w+)/updates', views.UpdateViewSet, base_name='Update')
+router.register(r'goals', views.GoalViewSet, base_name='Goal')
+router.register(r'^goals/(?P<goal_id>\w+)/plans', views.PlanViewSet, base_name='Plan')
+router.register(r'^stepOccurrences/(?P<stepOccurrence_id>\w+)/updateOccurrences', views.UpdateOccurrenceViewSet, base_name='UpdateOccurrence')
+router.register(r'^myProfile', views.PersonalProfileViewSet)
+
+router.register(r'^profiles', views.ProfileViewSet, base_name='Profile')
+router.register(r'^plan/search', views.PlanSearchViewSet, base_name="plan-search")
+
+
+
+from django.views.generic.base import RedirectView
+
 
 
 oauth2_endpoint_views = [
@@ -65,22 +89,25 @@ if settings.DEBUG:
     ]
 
 urlpatterns = [
-    url(r'^interest', views.interest, name="interest"),
+      url(r'^$', TemplateView.as_view(template_name='index.html')),
+      url(r'^interest', views.interest, name="interest"),
       url(r'^swagger/', schema_view),
       url(r'^api/hello', ApiEndpoint.as_view()),  # an example resource endpoint
       url(r'^api/', include(router.urls)),
       url(r'^o/', include('oauth2_provider.urls', namespace='oauth2_provider')),
+      url(r'^search/', include('haystack.urls')),
 
 
                 # url(r'^$', views.splash, name='splash'),
 
-      url(r'^$', TemplateView.as_view(template_name='index.html')),
       url(r'^admin/', admin.site.urls),
       url(r'^accounts/', include('allauth.urls')),
       url(r'^goals/add', views.goals_add, name='goals_add'),
       url(r'^secret', views.secret_page, name='secret'),
+      url(r'^tinymce/', include('tinymce.urls')),
+      url(r'^signS3Upload/$', views.sign_s3_upload, name='sign_s3_upload'),
 
-                # REST FRAMEWORK URLS
+                  # REST FRAMEWORK URLS
       # url(r'^users/$', views.UserList.as_view()),
       # url(r'^users/(?P<pk>[0-9]+)/$', views.UserDetail.as_view()),
       # url(r'^goals/$', views.GoalList.as_view()),
@@ -99,9 +126,6 @@ urlpatterns = [
       url(r'^rates/$', views.RateList.as_view()),
       url(r'^rates/(?P<pk>[0-9]+)$', views.RateDetail.as_view()),
 
-      url(r'^sessions/$', views.SessionList.as_view()),
-      url(r'^sessions/(?P<pk>[0-9]+)$', views.SessionDetail.as_view()),
-
       url(r'^updates/$', views.UpdateList.as_view()),
       url(r'^updates/(?P<pk>[0-9]+)$', views.UpdateDetail.as_view()),
 
@@ -114,6 +138,7 @@ urlpatterns = [
       url(r'^goals/(?P<goal_id>\w+)/plans', views.goals_plans, name='goals_plans'),
       url(r'^api-auth/', include('rest_framework.urls')),
       url(r'^plans/(?P<plan_id>\w+)', views.plan_edit, name='plan_edit'),
+      url(r'^favicon.ico$', RedirectView.as_view(url='/static/favicon2.ico'), name="favicon"),
 
 ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
 
