@@ -6,7 +6,7 @@ from django.core.urlresolvers import reverse_lazy
 from django.contrib.auth.decorators import login_required
 from django.http import HttpResponseRedirect
 from django.contrib.auth.models import User
-from kiterope.models import Goal, Plan, Step, Coach, Student, Session, Review, Update, Rate, Question, Answer, Interest
+from kiterope.models import Goal, Plan, Step, Coach, SearchQuery, Participant, Notification, Student, Session, Review, Profile, Update, Rate, Question, Answer, Interest, StepOccurrence, PlanOccurrence, UpdateOccurrence, UpdateOccurrenceManager, StepOccurrenceManager
 import datetime, time
 from time import mktime
 from datetime import datetime
@@ -17,11 +17,23 @@ from oauth2_provider.views.generic import ProtectedResourceView
 from django.http import HttpResponse
 from oauth2_provider.ext.rest_framework import TokenHasReadWriteScope, TokenHasScope
 
+from rest_framework.parsers import FormParser, MultiPartParser, JSONParser
+from rest_framework.decorators import detail_route
+from rest_framework.decorators import list_route
+
+from datetime import date
+from dateutil.rrule import rrule, DAILY
+
+from kiterope.permissions import IsAuthorOrReadOnly, AllAccessPostingOrAdminAll, IsOwnerOrReadOnly
 
 
 import requests
 import json
+import datetime
+from datetime import timedelta
 
+import boto
+import mimetypes
 
 
 from kiterope.helpers import formattime
@@ -29,9 +41,14 @@ from kiterope.helpers import formattime
 
 from rest_framework import viewsets
 from rest_framework.views import APIView
-from kiterope.serializers import UserSerializer, GoalSerializer, PlanSerializer, StepSerializer, CoachSerializer, QuestionSerializer
-from kiterope.serializers import StudentSerializer, SessionSerializer, ReviewSerializer, UpdateSerializer, RateSerializer, AnswerSerializer, InterestSerializer
+from kiterope.serializers import UserSerializer, SearchQuerySerializer, NotificationSerializer, UpdateOccurrenceSerializer, UpdateSerializer, ProfileSerializer, GoalSerializer, PlanSerializer, StepSerializer2, CoachSerializer, QuestionSerializer, StepOccurrenceSerializer, PlanOccurrenceSerializer
+from kiterope.serializers import StudentSerializer, SessionSerializer, ReviewSerializer, UpdateSerializer, PlanSearchSerializer, RateSerializer, AnswerSerializer, InterestSerializer
 from kiterope.forms import UserForm, ProfileForm, GoalForm, PlanForm, InterestForm
+from drf_haystack.viewsets import HaystackViewSet
+from drf_haystack.serializers import HaystackSerializer
+
+from kiterope.search_indexes import PlanIndex
+
 
 from rest_framework import status
 from rest_framework.decorators import api_view
@@ -42,8 +59,17 @@ from rest_framework import generics
 from rest_framework.decorators import api_view, renderer_classes
 from rest_framework import response, schemas
 from rest_framework_swagger.renderers import OpenAPIRenderer, SwaggerUIRenderer
+from rest_framework.permissions import IsAuthenticated
 
 
+# OpenTok is the protocol for Video and Voice chat that Kiterope uses
+from opentok import OpenTok, MediaModes, OpenTokException, __version__
+
+
+
+
+OPENTOK_API_KEY = "45757612"       # Replace with your OpenTok API key.
+OPENTOK_API_SECRET  = "a2287c760107dbe1758d5bc9655ceb7135184cf9"
 
 class ApiEndpoint(ProtectedResourceView):
     def get(self, request, *args, **kwargs):
@@ -59,10 +85,20 @@ class UserViewSet(viewsets.ModelViewSet):
     queryset = User.objects.all().order_by('-date_joined')
     serializer_class = UserSerializer
 
+
+
 class GoalViewSet(viewsets.ModelViewSet):
-    permission_classes = [permissions.IsAuthenticated, TokenHasScope]
-    required_scopes = ['groups']
+    model = Goal
+
     queryset = Goal.objects.all()
+
+    #permission_classes = [permissions.IsAuthenticated, TokenScope]
+    #authentication_classes = []
+
+    required_scopes = ['groups']
+
+    #parser_classes = (JSONParser, )
+
     serializer_class = GoalSerializer
 
     def create(self, request, *args, **kwargs):
@@ -71,6 +107,60 @@ class GoalViewSet(viewsets.ModelViewSet):
         self.perform_create(serializer)
         headers = self.get_success_headers(serializer.data)
         return Response(serializer.data, status=status.HTTP_201_CREATED, headers=headers)
+
+    def list(self, request, *args, **kwargs):
+        queryset = self.filter_queryset(self.get_queryset())
+
+        page = self.paginate_queryset(queryset)
+        if page is not None:
+            serializer = self.get_serializer(page, many=True)
+            return self.get_paginated_response(serializer.data)
+
+        serializer = self.get_serializer(queryset, many=True)
+        return Response(serializer.data)
+
+    def post(self, request, *args, **kwargs):
+        self.create(request, *args, **kwargs)
+        return self.list(request, *args, **kwargs)
+
+
+class StepViewSet(viewsets.ModelViewSet):
+    #permission_classes = [permissions.IsAuthenticated, TokenHasScope]
+    #required_scopes = ['groups']
+
+    serializer_class = StepSerializer2
+    queryset = Step.objects.all()
+    required_scopes = ['groups']
+
+    def get_queryset(self):
+
+        try:
+            plan_id = self.kwargs['plan_id']
+            print(plan_id)
+            aQueryset = Step.objects.filter(plan=plan_id)
+        except:
+            aQueryset = Step.objects.all()
+
+        return aQueryset
+
+    def create(self, request, *args, **kwargs):
+        print(request.data)
+        serializer = self.get_serializer(data=request.data)
+
+        serializer.is_valid(raise_exception=True)
+
+        self.perform_create(serializer)
+        headers = self.get_success_headers(serializer.data)
+        return Response(serializer.data, status=status.HTTP_201_CREATED, headers=headers)
+
+    def update(self, request, *args, **kwargs):
+        instance = self.get_object()
+
+        serializer = self.get_serializer(instance, data=request.data)
+        serializer.is_valid(raise_exception=True)
+        self.perform_update(serializer)
+
+        return Response(serializer.data)
 
 
     def list(self, request, *args, **kwargs):
@@ -86,19 +176,542 @@ class GoalViewSet(viewsets.ModelViewSet):
 
 
     def post(self, request, *args, **kwargs):
+        print(request)
+
         self.create(request, *args, **kwargs)
         return self.list(request, *args, **kwargs)
 
-class PlanViewSet(viewsets.ModelViewSet):
-    queryset = Plan.objects.all()
-    serializer_class = PlanSerializer
+    @list_route(methods=['get'], permission_classes=[permissions.IsAuthenticated], url_path='daily')
+    def dailyList(self, request, *args, **kwargs):
+        print("inside")
 
-class StepViewSet(viewsets.ModelViewSet):
-    permission_classes = [permissions.IsAuthenticated, TokenHasScope]
+class PlanOccurrenceViewSet(viewsets.ModelViewSet):
+
+    serializer_class = PlanOccurrenceSerializer
+
+
+class UpdateOccurrenceViewSet(viewsets.ModelViewSet):
+    queryset = UpdateOccurrence.objects.all()
     required_scopes = ['groups']
-    queryset = Step.objects.all()
-    serializer_class = StepSerializer
+    serializer_class = UpdateOccurrenceSerializer
 
+    def get_queryset(self):
+
+        try:
+            stepOccurrence_id = self.kwargs['stepOccurrence_id']
+            aQueryset = UpdateOccurrence.objects.filter(stepOccurrence=stepOccurrence_id)
+            print("aQueryset %s" % aQueryset.count())
+        except:
+            aQueryset = UpdateOccurrence.objects.all()
+
+        return aQueryset
+
+    def list(self, request, *args, **kwargs):
+        queryset = self.filter_queryset(self.get_queryset())
+
+        page = self.paginate_queryset(queryset)
+        if page is not None:
+            serializer = self.get_serializer(page, many=True)
+            return self.get_paginated_response(serializer.data)
+
+        serializer = self.get_serializer(queryset, many=True)
+        return Response(serializer.data)
+
+
+
+class UpdateViewSet(viewsets.ModelViewSet):
+    queryset = Update.objects.all()
+    required_scopes = ['groups']
+    serializer_class = UpdateSerializer
+
+    def get_queryset(self):
+
+        try:
+            step_id = self.kwargs['step_id']
+            aQueryset = Update.objects.filter(step=step_id)
+        except:
+            aQueryset = Update.objects.all()
+
+        return aQueryset
+
+    def create(self, request, *args, **kwargs):
+        print(request.data)
+        serializer = self.get_serializer(data=request.data)
+
+        serializer.is_valid(raise_exception=True)
+
+        self.perform_create(serializer)
+        headers = self.get_success_headers(serializer.data)
+        return Response(serializer.data, status=status.HTTP_201_CREATED, headers=headers)
+
+    def update(self, request, *args, **kwargs):
+        instance = self.get_object()
+
+        serializer = self.get_serializer(instance, data=request.data)
+        serializer.is_valid(raise_exception=True)
+        self.perform_update(serializer)
+
+        return Response(serializer.data)
+
+
+    def list(self, request, *args, **kwargs):
+        queryset = self.filter_queryset(self.get_queryset())
+
+        page = self.paginate_queryset(queryset)
+        if page is not None:
+            serializer = self.get_serializer(page, many=True)
+            return self.get_paginated_response(serializer.data)
+
+        serializer = self.get_serializer(queryset, many=True)
+        return Response(serializer.data)
+
+
+    def post(self, request, *args, **kwargs):
+        print(request)
+
+        self.create(request, *args, **kwargs)
+        return self.list(request, *args, **kwargs)
+
+
+
+class StepOccurrenceViewSet(viewsets.ModelViewSet):
+    permission_classes = [permissions.IsAuthenticated, TokenHasScope]
+    #required_scopes = ['groups']
+    stepStart = -999
+    stepEnd = -999
+    serializer_class = StepOccurrenceSerializer
+
+    def list(self, request, *args, **kwargs):
+
+        try:
+
+            periodRangeStart = self.kwargs['periodRangeStart']
+            periodRangeEnd = self.kwargs['periodRangeEnd']
+
+            #print("inside list2")
+            periodRangeStart = datetime.date.strptime(periodRangeStart, "%Y-%m-%d")
+            periodRangeEnd = datetime.date.strptime(periodRangeEnd, "%Y-%m-%d")
+            print("periodRangeStart %s" % periodRangeStart)
+            print("periodRangeStart %s" % periodRangeEnd)
+
+        except:
+            #periodRangeStart = str(datetime.datetime.now().date())
+            periodRangeStart = datetime.datetime.now().date() - datetime.timedelta(days=1) #+ datetime.timedelta(days=10)
+
+            periodRangeEnd = periodRangeStart
+            #print ("periodRangeStart " + periodRangeStart)
+            #print ("periodRangeEnd " + periodRangeEnd)
+
+        currentUser = request.user
+
+
+
+        self.updateOccurrences(currentUser, periodRangeStart, periodRangeEnd)
+        print("occurrencesUpdated")
+
+        queryset = self.filter_queryset(self.get_queryset())
+        page = self.paginate_queryset(queryset)
+        if page is not None:
+            serializer = self.get_serializer(page, many=True)
+            return self.get_paginated_response(serializer.data)
+
+
+        serializer = self.get_serializer(queryset, many=True)
+        return Response(serializer.data)
+
+
+
+
+    def post(self, request, *args, **kwargs):
+        print(request)
+
+        self.create(request, *args, **kwargs)
+        return self.list(request, *args, **kwargs)
+
+    def updateOccurrences(self, currentUser, periodRangeStart, periodRangeEnd):
+        print("updateOccurrences called")
+        try:
+            print("currentUser %s" % currentUser)
+            userPlanOccurrences = PlanOccurrence.objects.filter(user=currentUser.id)
+            print("userPlanOccurrences %d" % userPlanOccurrences.count())
+
+            print("before persisted occurrences")
+
+            persistedOccurrences = StepOccurrence.objects.filter(user=currentUser.id)
+
+            print("persistedOccurrences %d" % persistedOccurrences.count())
+
+
+
+            for aPlanOccurrence in userPlanOccurrences:
+                #print("aPlanOccurrence.startDate %s" % type(aPlanOccurrence.startDate))
+                #periodRangeStart = datetime.datetime.strftime(periodRangeStart, '%Y-%m-%d')
+                #print("periodRangeStart %s" % type(periodRangeStart))
+
+                #planOccurrenceStartDate = datetime.datetime.strftime(aPlanOccurrence.startDate, '%Y-%m-%d')
+
+                periodStart = periodRangeStart - aPlanOccurrence.startDate
+
+                periodStart = periodStart.days
+                print("periodStart %s" % periodStart)
+
+
+                periodEnd = periodRangeEnd - aPlanOccurrence.startDate
+                periodEnd = periodEnd.days
+                print("periodEnd %s" % periodEnd)
+
+
+
+
+                #print("aPlanOccurrence.plan %d" % aPlanOccurrence.plan.id)
+
+                thePlan = Plan.objects.filter(id=aPlanOccurrence.plan.id)
+                print("thePlan %d" % thePlan.count())
+
+                #print("beforePlansSteps")
+                thePlansSteps = Step.objects.filter(plan=thePlan[0].id)
+                print("thePlansSteps %d" % thePlansSteps.count())
+
+                for aStep in thePlansSteps:
+                    print("inside thePlansSteps loop")
+
+                    #print("inside comparePeriodToStep")
+                    print("aStep.frequency %s" % aStep.frequency)
+                    print("aStep.id %d" % aStep.id)
+
+                    if (aStep.frequency == "ONCE"):
+                        #print("aStep.startDate %s" % aStep.startDate)
+
+                        stepStart = aStep.startDate
+                        stepEnd = aStep.startDate
+
+
+
+                    elif (aStep.frequency == "DAILY"):
+                        #print("aStep.startDate %s" % aStep.startDate)
+                        #print("aStep.endDate %s" % aStep.endDate)
+
+                        stepStart = aStep.startDate
+                        stepEnd = aStep.endDate
+
+                    elif (aStep.frequency == "WEEKLY"):
+                        daysList = [aStep.day01, aStep.day02, aStep.day03, aStep.day04, aStep.day05, aStep.day06,
+                                    aStep.day07]
+                        #print("daysList %s" % aStep.day01, aStep.day02, aStep.day03, aStep.day04, aStep.day05, aStep.day06,
+                                    #aStep.day07)
+                        try:
+                            aStepStart = daysList.index(True)
+                            print ("aStepStart %s" % aStepStart)
+
+                            aStepEnd = len(daysList) - daysList[::-1].index(True) - 1
+                            stepStart = ((aStep.begins - 1) * 7) + aStepStart
+                            stepEnd = ((aStep.ends - 1) * 7) + aStepEnd
+                        except:
+                            pass
+
+
+                    elif (aStep.frequency == "MONTHLY"):
+                        print("frequency = monthly - work needs to be done")
+
+                    # stepStart = aPlanOccurrence.startDate + datetime.timedelta(days=stepStart)
+                    # stepEnd = aPlanOccurrence.startDate + datetime.timedelta(days=stepEnd)
+                    #print("stepStart %d" % stepStart)
+                    #print("stepEnd %d" % stepEnd)
+                    #print("periodStart %d" % periodStart)
+                    #print("periodEnd %d" % periodEnd)
+                    print("aPlanOccurence startDate %s" % aPlanOccurrence.startDate)
+
+                    if (stepStart <= periodStart & stepEnd >= periodEnd):
+                        print("periodStart to periodEnd")
+
+                        iterationStart = aPlanOccurrence.startDate + datetime.timedelta(days=periodStart)
+                        iterationEnd = aPlanOccurrence.startDate + datetime.timedelta(days=periodEnd)
+
+                    elif (stepStart <= periodStart & stepEnd <= periodEnd):
+                        print("periodStart to StepEnd")
+
+                        iterationStart =  aPlanOccurrence.startDate + datetime.timedelta(days=periodStart)
+                        iterationEnd = aPlanOccurrence.startDate + datetime.timedelta(days=stepEnd)
+
+
+
+                    elif (stepStart >= periodStart & stepEnd <= periodEnd):
+                        print("stepStart to stepEnd")
+
+                        #print("inside comparison =2")
+
+                        iterationStart = aPlanOccurrence.startDate + datetime.timedelta(days=stepStart)
+                        #print("iterationStart %s" % iterationStart)
+
+                        iterationEnd = aPlanOccurrence.startDate + datetime.timedelta(days=stepEnd)
+                        #print("iterationEnd %s" % iterationEnd)
+
+                    elif (stepStart <= periodStart & stepEnd > periodEnd):
+                        print("stepStart to periodEnd")
+
+                        iterationStart = aPlanOccurrence.startDate + datetime.timedelta(days=stepStart)
+                        iterationEnd = aPlanOccurrence.startDate + datetime.timedelta(days=periodEnd)
+                    else:
+                        pass
+                    print("periodStart %s" % periodStart)
+                    print("periodEnd %s" % periodEnd)
+
+                    print("stepStart %s" % stepStart)
+                    print("stepStart %s" % stepEnd)
+
+
+                    stepOccurrenceExists = False
+
+                    #iterationStart = datetime.datetime.strptime(iterationStart, '%Y-%m-%d')
+                    #iterationEnd = datetime.datetime.strptime(iterationEnd, '%Y-%m-%d')
+
+                    print("before DateIterator loop")
+
+                    for dateIterator in rrule(DAILY, dtstart=iterationStart, until=iterationEnd):
+                    #for dateIterator in range(iterationStart, iterationEnd):
+                        print("inside dateIteration")
+
+
+                        #print("inside false stepOccurrence loop")
+                        #print("aStep.id %d" % aStep.id)
+                        #print("dateIterator %s" % dateIterator)
+                        #print("aPlanOccurrence %d" % aPlanOccurrence.id)
+                        #print("currentUser %d" % currentUser.id)
+
+                        #print("aStepOccurrence ")
+                        for persistentOccurrence in persistedOccurrences:
+                            #print("currentPersistentOccurence.step.id %s" % persistentOccurrence.step.id)
+                            #print("aStep.id %s" % aStep.id)
+
+                            if aStep.id == persistentOccurrence.step.id:
+                                print("dateIterator is %s" % type(dateIterator))
+                                print("persistentOccurrence.date is %s" % type(persistentOccurrence.date))
+                                print("persistentOccurrence.date %s" % persistentOccurrence.date)
+                                print("dateIterator %s" % dateIterator)
+                                print("inside aStep.id == persistentOccurrence.step.id")
+
+                                if dateIterator.date() == persistentOccurrence.date.date():
+                                    print("inside dateIterator == persistentOccurrence.date")
+
+                                    print("persistentOccurrence.planOccurrence.id %d" % persistentOccurrence.planOccurrence.id)
+                                    print("aPlanOccurrence.id %s" % aPlanOccurrence.id)
+                                    if aPlanOccurrence.id == persistentOccurrence.planOccurrence.id:
+                                        print("inside aPlanOccurrence.id == persistentOccurrent.id")
+                                        stepOccurrenceExists=True
+                                        print("stepOccurrenceExists %s" % stepOccurrenceExists)
+
+                                        break
+                        print("stepOccurrenceExists %s" % stepOccurrenceExists)
+                        if (stepOccurrenceExists == False):
+                            aStepOccurrence = StepOccurrence.objects.create_occurrence(aStep.id, dateIterator,
+                                                                                       aPlanOccurrence.id,
+                                                                                       currentUser.id)
+                            print("aStep.id = %s" % aStep.id )
+                            currentStepUpdates = Update.objects.filter(step=aStep.id)
+
+                            for currentStepUpdate in currentStepUpdates:
+                                print("inside currentStepUpdates")
+                                anUpdateOccurrence = UpdateOccurrence.objects.create_occurrence(aStepOccurrence.id, currentStepUpdate.id)
+
+                                print("Update create Occurrence finished")
+
+                            print("stepOccurrence Doesn't Exist")
+
+                            #aStepOccurrence.save()
+
+
+
+        except:
+            pass
+
+
+
+    def get_queryset(self):
+        # user = self.request.user
+        # userPlanQueryset = user.plan_set.all()
+        try:
+            print("inside Get Queryset")
+            periodRangeStart = self.kwargs['periodRangeStart']
+            periodRangeEnd = self.kwargs['periodRangeEnd']
+            print("inside Get Queryset2")
+
+            #periodRangeStart = datetime.date.strptime(periodRangeStart, "%Y-%m-%d")
+            #periodRangeEnd = datetime.date.strptime(periodRangeEnd, "%Y-%m-%d")
+
+            print("getQueryset periodRangeStart %s periodRangeEnd" % (periodRangeStart, periodRangeEnd))
+
+
+        except:
+            #periodRangeStart = str(datetime.datetime.now().date())
+            periodRangeStart = datetime.datetime.now().date()
+
+            periodRangeEnd = periodRangeStart
+
+
+        theQueryset = StepOccurrence.objects.all()
+
+        return theQueryset
+
+
+    def comparePeriodToStep(self, periodStart, periodEnd, aStep, aPlanOccurrence):
+
+        return range(iterationStart, iterationEnd)
+
+class NotificationViewSet(viewsets.ModelViewSet):
+    serializer_class = NotificationSerializer
+    permission_classes = [IsOwnerOrReadOnly]
+
+    def get_queryset(self):
+        theUser = self.request.user
+        return Notification.objects.filter(user_id=theUser.id)
+
+
+
+
+
+class PlanViewSet(viewsets.ModelViewSet):
+    serializer_class = PlanSerializer
+    queryset = Plan.objects.all()
+    permission_classes = [IsAuthorOrReadOnly]
+
+    required_scopes = ['groups']
+
+    def update(self, request, *args, **kwargs):
+        instance = self.get_object()
+
+        serializer = self.get_serializer(instance, data=request.data)
+        serializer.is_valid(raise_exception=True)
+        self.perform_update(serializer)
+
+        return Response(serializer.data)
+
+class ProfileViewSet(viewsets.ModelViewSet):
+    serializer_class = ProfileSerializer
+    queryset = Profile.objects.all()
+    permission_classes = [IsOwnerOrReadOnly]
+    required_scopes = ['groups']
+
+
+
+
+
+
+
+class SessionViewSet(viewsets.ModelViewSet):
+
+    serializer_class = SessionSerializer
+    queryset = Session.objects.all()
+    permission_classes = [IsAuthenticated]
+    required_scopes = ['groups']
+
+    def create(self, request, *args, **kwargs):
+        OPENTOK_API_KEY = "45757612"  # Replace with your OpenTok API key.
+        OPENTOK_API_SECRET = "a2287c760107dbe1758d5bc9655ceb7135184cf9"
+
+        opentok_sdk = OpenTok(OPENTOK_API_KEY, OPENTOK_API_SECRET)
+
+
+
+
+        serializer = self.get_serializer(data=request.data)
+        userProfileBeingCalledId = request.data['userProfileBeingCalledId']
+        session = opentok_sdk.create_session()
+
+        try:
+
+            theTokBoxSessionId = session.session_id
+            theTokBoxToken = session.generate_token()
+
+
+            theSession = Session.objects.create_session(theTokBoxSessionId, theTokBoxToken)
+
+            print("theSession")
+
+
+            print("theSession.id %s" % theSession.id)
+            primaryParticipant = Participant.objects.create_participant(request.user.id, True, theSession.id)
+
+            theUserBeingCalled = Profile.objects.filter(id=userProfileBeingCalledId).first()
+
+            secondaryParticipant = Participant.objects.create_participant(theUserBeingCalled.user_id, False, theSession.id)
+            notification = Notification.objects.create_notification(theUserBeingCalled.user_id, theSession.id, "CALL")
+            notification = Notification.objects.create_notification(request.user.id, theSession.id, "CALL")
+
+
+
+
+        except:
+            pass
+
+        serializer.is_valid(raise_exception=True)
+
+        print("this is the serializer data %s" % serializer.data)
+
+        #self.perform_create(serializer)
+        headers = self.get_success_headers(serializer.data)
+        return Response(serializer.data, status=status.HTTP_201_CREATED, headers=headers)
+
+
+class SearchQueryViewSet(viewsets.ModelViewSet):
+    serializer_class = SearchQuerySerializer
+    queryset = SearchQuery.objects.all()
+    permission_classes = [AllAccessPostingOrAdminAll]
+
+
+
+
+
+class PlanSearchViewSet(HaystackViewSet):
+    index_models = [Plan]
+    serializer_class = PlanSearchSerializer
+    permission_classes = [IsAuthorOrReadOnly]
+
+
+
+
+
+
+
+class PersonalProfileViewSet(viewsets.ModelViewSet):
+    serializer_class = ProfileSerializer
+    required_scopes = ['groups']
+    permission_classes = [IsOwnerOrReadOnly]
+    queryset = Profile.objects.all()
+
+
+    def list(self, request, *args, **kwargs):
+        queryset = self.filter_queryset(self.get_queryset())
+
+        page = self.paginate_queryset(queryset)
+        if page is not None:
+            serializer = self.get_serializer(page, many=True)
+            return self.get_paginated_response(serializer.data)
+
+        serializer = self.get_serializer(queryset, many=True)
+        return Response(serializer.data)
+
+
+    def update(self, request, *args, **kwargs):
+        instance = self.get_object()
+
+        serializer = self.get_serializer(instance, data=request.data)
+        serializer.is_valid(raise_exception=True)
+        self.perform_update(serializer)
+
+        return Response(serializer.data)
+
+
+    def get_queryset(self):
+
+
+
+        """
+        This view should return a list of all the purchases
+        for the currently authenticated user.
+        """
+        theUser = self.request.user
+
+        return Profile.objects.filter(user_id=theUser.id)
 
 
 
@@ -168,18 +781,6 @@ class PlanDetail(generics.RetrieveUpdateDestroyAPIView):
     serializer_class = PlanSerializer
 
 
-class StepList(generics.ListCreateAPIView):
-    permission_classes = [permissions.IsAuthenticated, TokenHasScope]
-    required_scopes = ['groups']
-    queryset = Step.objects.all()
-    serializer_class = StepSerializer
-
-
-class StepDetail(generics.RetrieveUpdateDestroyAPIView):
-    permission_classes = [permissions.IsAuthenticated, TokenHasScope]
-    required_scopes = ['groups']
-    queryset = Step.objects.all()
-    serializer_class = StepSerializer
 
 
 
@@ -231,14 +832,7 @@ class ReviewDetail(generics.RetrieveUpdateDestroyAPIView):
     serializer_class = ReviewSerializer
 
 
-class SessionList(generics.ListCreateAPIView):
-    queryset = Session.objects.all()
-    serializer_class = SessionSerializer
 
-
-class SessionDetail(generics.RetrieveUpdateDestroyAPIView):
-    queryset = Session.objects.all()
-    serializer_class = SessionSerializer
 
 
 
@@ -427,7 +1021,7 @@ def plan_detail(request, pk, format=None):
         plan.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
-
+'''
 @api_view(['GET', 'PUT', 'DELETE'])
 def steps_list(request, format=None):
     try:
@@ -475,7 +1069,7 @@ def step_detail(request, pk, format=None):
     elif request.method == 'DELETE':
         step.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
-
+'''
 @login_required  
 def goals_add(request):    
     if request.method=='POST':
@@ -550,7 +1144,23 @@ def plan_edit(request, plan_id=None):
 
     
     
-    
+conn = boto.connect_s3(settings.AWS_ACCESS_KEY_ID, settings.AWS_SECRET_ACCESS_KEY)
+
+def sign_s3_upload(request):
+    print("this is the request %s" % request)
+    object_name = request.GET['objectName']
+    content_type = mimetypes.guess_type(object_name)[0]
+
+    signed_url = conn.generate_url(
+        300,
+        "PUT",
+        settings.AWS_STORAGE_BUCKET_NAME,
+        'images/' + object_name,
+        headers = {'Content-Type': content_type,
+                   'x-amz-acl': 'public-read',
+                   })
+
+    return HttpResponse(json.dumps({'signedUrl': signed_url}))
     
     
     
