@@ -1,6 +1,6 @@
-var path = require("path")
-var webpack = require('webpack')
-var BundleTracker = require('webpack-bundle-tracker')
+var path = require("path");
+var webpack = require('webpack');
+var BundleTracker = require('webpack-bundle-tracker');
 
 
 module.exports = {
@@ -19,15 +19,21 @@ module.exports = {
     new webpack.ProvidePlugin({
       $: 'jquery',
       jQuery: 'jquery',
-      'window.jQuery': 'jquery'
+      'window.jQuery': 'jquery',
+      'window.$':'jquery',
     })
   ],
 
   module: {
     loaders: [
-      { test: /\.jsx?$/, exclude: /node_modules/, loader: 'babel-loader', query:
+      { test: /\.jsx?$/,
+        exclude: /node_modules/,
+        loader: 'babel-loader',
+        query:
       {
-        presets:['react','es2015']
+        presets:['react','es2015',  'stage-2'],
+        plugins: ["transform-decorators-legacy", "transform-class-properties", ]
+
       }},
       { test: /\.css$/, loader: "style-loader!css-loader" }
 // to transform JSX into JS
@@ -38,13 +44,17 @@ module.exports = {
     modulesDirectories: ['node_modules', 'bower_components', ],
     extensions: ['', '.js', '.jsx'],
     fallback: path.join(__dirname, "node_modules"),
+    alias: {
+      react: path.resolve('./node_modules/react'),
+      'react/lib/ReactMount': 'react-dom/lib/ReactMount' }
+
   },
 
   resolveLoader: { fallback: path.join(__dirname, "node_modules") }
 
-}
+};
 WEBPACK_LOADER = {
     'DEFAULT': {
         'STATS_FILE': path.join(__dirname, 'webpack-stats.json')
     }
-}
+};
